Allow the SockJS server to connect to WebSocket clients.
This commit allows the SockJS server to connect to WebSocket clients on the url http://site:port+1/sockjs/websocket.

I understand that there are issues with websockets on the client, so I've left that disabled, but I'm not sure why they were disabled on the server. This change allows non-browsers to easily interact with the meteor server. I'm working on an implementation of DDP for other languages and this is the only way that sockjs supports non-browser clients speaking to it.

If there is a compelling reason to disable websockets server-side, then I'll try to go through the rigamarole of reimplementing SockJS's protocol outside of the browser..