Changed madewith.meteor.com/ link generated for apps hosted on a 2-level subdomain of meteor.com
For example, the badge on foo.bar.meteor.com now points to
madewith.meteor.com/foo.bar.meteor.com. I updated comments on the madewith
app in a separate commit to reflect this.

This was tested by deploying a sample app to ab.meteor.com and a.b.meteor.com
and verifying that the links go to the right URL. Upvoting was not tested
explicitly but the code uses the same url for that.