Unify `template` obj and template callback `this`.
The `template` arg in `eventHandler(event, template)` and the `this`
in all three of {create,render,destroy} are now an object with
methods `find`,`findAll` assigned at creation time (closures) and
properties `firstNode`, `lastNode`, and `data` (the handlebars data)
set upon each callback.  No other props are set and the app is free
to scribble on this object.

Only subtlety is that we can't support find/findAll/firstNode/lastNode
in create/destroy callback.  In this case, find/findAll throw an error
and firstNode/lastNode are reliably null.

Added landmark.hasDom().