Make Meteor.Cursor.forEach fully synchronous even if the user's callback yields.
Previously, if the callback yields (eg, if it does a DB query), the next
iteration of the callback could run, or the forEach could finish entirely,
because the future we waited on only checked to ensure that the callback was
*started* on all documents, not *finished*.

Fixes #321. Thanks to Scott Anderson for bug report with minimal reproduction
case.