Fix race condition in spiderable.
If the "check to see if page is ready" interval ran when "Meteor" is defined but
"Meteor.status" is not yet defined, an error message about trying to call an
undefined function would be dumped into the output source before the page
itself.

This can be easily reproduced by changing the timer interval from 100 to 1.