Fix #353: livedata_connection state corruption with "wait: true" methods.
When a "wait: true" method is ready to fire its callback, and there are blocked
methods but no blocked "wait" methods, we used to leave the blocked methods in
the blocked_methods array in addition to sending them. In addition to leading to
potentially leading to duplicate method calls (ignored by the server), this
blocked all code push migrations.