Clarified when reactivity works
After reading the Reactivity section for the first time,
I was under the slight impression that a reactive context would
be sufficient for triggering changes. E.g. if the name in the
fragment example were set not from Session, but from a global,
it could be surmised that the HTML fragment would still be
automatically updated if the global variable changed.

This small patch tries to make it clearer that reactivity requires
*both* a reactive context and a reactive data source.