Fixed a problem with deep object selectors with differing keys in minimongo.
Select from minimongo would match objects who different in deep keys (not
values). It seems like the object matching of selector.js was intended to work
this way (the name `b_keys` implies that it should be the keys of `b` after
all), but was mis-implemented.

Fixes #455 - a subtle bug with almost-overlapping subscriptions

(glasser added another test case and cleaned up test whitespace.)