Refactor mongo-livedata cursor objects.
Separate the "description" of the cursor from the actual wrapped Mongo
cursor.

Collection.find() no longer runs a Mongo query: you have to actually call a
function like fetch, count, or observe on the cursor to run the
query. (Specifically, this lets us skip a useless query when a publish function
returns the result of a find().) This implies that errors that used to be
reported at find() time are now reported later.

Also, allow objects other than Mongo cursors returned from publish functions to
define their own _publishCursor functions (this is not documented or officially
supported). The _publishCursor code is moved from the livedata package to
mongo-livedata.