Added a fast path to compilation of selectors if only the _id of a document is queried.
In my application, quite some time is spent in the _compileSelector function.
I assume the 'eval' to be the main culprit (it's hard to profile exactly, though).
A good deal of all queries, however, are pretty simple, so I added a fast path
for the simplest case, which in my case is almost half of all the queries my
application issues.