Better management of user documents when created from external service.
Don't lose fields that were added in a previous login, but that aren't
received on a subsequent one. An example of this is the refresh token
sent from Google only on first login, in the case that the service
is configured on "offline" mode.

This is a prerequisite for merging in #464