Add extraFields config option to Accounts.ui
Allows you to define extra fields for Accounts.ui signup form.

The fields must have a fieldName and fieldLabel attributes and can have
visible and validate functions.

The info introduced in those fields is putted into the account
profile's data using fieldName as attribute's name.

The validate functions are called with two arguments, value and
errorFn. Value contains the submitted value for that field and errorFn
is a function you can call with a custom error message as a string,
like errorFn('Name must have a value.');.