New facts package: a proof-of-concept implementation of server-side statistics.
Publishes statistics from the server to a Minimongo collection,
Meteor.Facts.server. (On the server side, Facts is *not* backed by MongoDB.) By
default, this is available to all users if autopublish is on, or no users if it
is not on, but you can configure this with Meteor.Facts.setUserIdFilter.

You can add an ugly view on the stats to your app with {{> serverFacts }}.

Current stats, by package:

- mongo-livedata:
  - observe-handles
  - live-results-sets (when this is less than observe-handles, cursor de-dup is helping you)
- livedata
  - subscriptions
  - crossbar-listeners
  - sessions-attached
  - sessions-detached
  - sessions (should be sum of sessions-attached and sessions-detached)