Upgrade useragent to 2.0.1.
Will be using for the appcache.  I tested 2.0.1 and it seems to work
fine.  I figured that all else being equal it would be good to use the
latest version to give us the most recent user agent matches, and so
that if we had any bug reports we wouldn't be out of date.