fix for oauth on mobile devices.
the problem with original oauth is that it wouldn't work for meteor inside phonegap because it will popup a window for auth, which won't work.
so right now, check whether on mobile. if on mobile, then do not popup a window, directly redirect to the url.
this method is found by boundsj, on this issue https://github.com/meteor/meteor/issues/438