Preserve tag filter state on reload in "todos" example.
When the "todos" example app reloads following a code change, the
current state of the tag filter is preserved by session
migration... but is then lost following the reload in two places: on
startup the tag filter state is unilaterally set to `null`, and is set
to `null` again when initializing the list_id.

This commit avoids the unilateral initialization, and only sets the
tag filter to `null` if the selected list has changed.