Experimental API for specifying "invalidation keys" when doing server-side Mongo reads and writes.
find(), update(), insert(), and remove() now all take an option named
"_invalidationKey". (For insert and remove, there's a new options argument just
to take this option.)

By default, the invalidation key for a read or write is the object

   {collection: COLLECTIONNAME}

If the query specifies a specific ID in the selector (or more generally, a list
of IDs with {_id: {$in: ["x", "y", "z"]}}), the default invalidation key also
contains the field id: ID.

If you specify an _invalidationKey option to any of these commands, all fields
in the given key will be added to the defaults.

So what is an invalidation key?  Well, as described here, all reads (well, all
observeChanges live reads) and all writes have an invalidation key.  If all of
the fields that exist in both a read's key and a write's key have the same
values, then the write triggers the read to refresh itself from the database. If
any field which exists in both keys has different values, it does not refresh.