Remove partial fix for connecting back to self in environments where absoluteUrl doesn't match.
There were more places where this change would be needed, and it was kinda ugly to try to share the code. Having absoluteUrl be where we are reachable is a requirement for testing OAuth flow anyway, having it required for automated testing isn't so bad.

This is typically only a problem when testing on local windows VMs. In this case, you can work around the issue by editing `C:\Windows\System32\drivers\etc\hosts` to include '10.0.2.2  localhost`.

(And yes, I know this code was my idea in the first place. I was wrong.)