`check` with predicate functions, introducing declareConstructor.
Allows predicates to be used as a match pattern, replacing
`Match.Where`.

Constructors can be used in a match pattern either with
`Match.InstanceOf`, or by making the constructor known to Meteor with
`Meteor.declareConstructor`.

Predicates are required to return a boolean `true` to match, which
avoids the problem of an undeclared constructor being used as a match
pattern silently succeeding for arbitrary input.