Stop livedata from depending on routepolicy.
That brought in all of webapp, which only works on "type: traditional" targets,
not server-only.

(We really should split the Node DDP Client code from the Node DDP Server code
instead.)

(Maybe routepolicy should just be part of webapp, or webapp should depend on
routepolicy? It's awkward to need to list it separately in various places...)