Prevent ctl from choosing a port at random.
When deploying an app to galaxy, localPort: 0 in the deployConfig was overriding
the port assigned by satellite to the server, so the server's port was being
chosen at random, so I removed localPort from the deployConfig. (Not 100% sure
this is the right fix.)
Deploying a simple app to galaxy works now, woohoo!