Fix a couple failing mongo-livedata tests.
Transform was being applied to every single Mongo query, including those issued
from _pollMongo. This led to the transformed doc being sent over DDP, and then
the client applied the transformation again. Instead, pass the useTransform flag
to prevent _pollMongo from transforming, so that the whole doc gets sent over
DDP and then the client does the transform, which is (I think?) what is supposed
to happen. I also think but am not sure that transform is supposed to be applied
for SynchronousCursors created for tailable cursors, so I had those
SynchronousCursors have useTransform=true.