Fix bug where we would never quiesce if we tried to revive subs that errored out.
This specifically would happen if you ran an app with at least one sub with a
ready message (ie not autopublish; eg parties) with auth, logged in, and then
killed the app and replaced it with a different app.  Hot code push would stall
and the new app would not load until you hit reload.

Why? Due to the login method call, hot code push had to wait for the login
method to be done. A method being done, in this case, includes having all its
data visible. This means waiting for the at-reconnect global quiescence to
finish. But this would never happen, since that requires waiting for all subs
that had been ready to become ready again, and instead the subs were getting
nosub because this is an entirely different app.