When removing the last NPM dependency, delete the NPM directory too.
Apply this to the current tree, which deletes the unused .npm directory from the
ctl package.

Specifically, this deletes the ".npm/package" or ".npm/plugin/foo" directory. We
don't attempt to also clean up the ".npm/plugin" or ".npm" directory, but a "git
clean -df" will do the trick if you really care.