Introduce concept of "build version" to unipackages.
We bump the "build version" of the tools precisely when we want to rebuild all
packages.  This handles the "I ran meteor update, I may need to rebuild my
app/$PACKAGE_DIRS packages" case, as well as ensuring that warehouse package
version numbers are incremented when we need to rebuild all packages (but not on
EVERY tools directory change).