Refactor argument parsing and 'help' display.
argumentParser is now required, and it takes in an optimist object instead of
needing to create one.

The name of the command is automatically removed from argv._ between
argumentParser and func, instead of having to deal with it in func.

We no longer are able to cleverly show the "-f" in the help for logs -f only if
talking to Galaxy. The logic there is too complex.

(The motivation for this commit is that "meteor help logs" and "meteor logs
--help" were both broken, in different ways.)