Remove "write to the global" coffeescript instructions.
As part of a docs pass we will explain the new way to use coffeescript globals.

(In short: in a package, anything declared with `api.export` becomes
package-level and exported. If you want something package-level and not
exported, or app-level, there's an object `share` and you can assign to fields
on it.)