Fall back to a different proxy service if the preferred one isn't available.
This is messy in many ways (among other things, "proxy" and "adminProxy"
shouldn't be hardcoded as the only two proxy service names), but it allows us to
run a galaxy on one satellite, by having apps bind to the adminProxy when the
normal proxy isn't running. This is also untested; running two satellites on the
same machine doesn't work because the two proxies will both try to bind to the
same discoveryPort.