Various bugs involving idStringify.
In minimongo, we support various types as _id fields; the most important are
strings and ObjectIds. We have a special encoding that we use when we need to
represent them as strings, but we had a few bugs with using it.

minimongo: The duplicate-ID check in insert needed to check using the encoded
string.

spark OrderedDict: due to linker changes, it was not successfully finding
LocalCollection._idStringify, and so it fell back to the identity
function. Fixed to use weak dependencies. Also, later refactorings have removed
the need for idStringify(null) === null so that special case is removed.

spark branch labels: use idStringify in various places where _id is used as a
branch label.