On the server, make the Mongo collection operation result available.
Some MongoDB design patterns such as
[Update if Current](http://docs.mongodb.org/manual/tutorial/isolate-sequence-of-operations/#update-if-current)
need access to the Mongo result.

This only makes sense on the server, since the client doesn't have a
way to do atomic updates by itself anyway.

For collection methods where a callback is included, call the callback
with the Mongo result as the third parameter.

(This means the Mongo result isn't available when using collection
methods synchronously, but this is a pretty specialized case and the
developer is more likely to be using a package providing e.g.
optimistic update-with-retry functionality rather than using it
directly).