start of html5 module as meteor package
It would be nice to just use this module to do spec-compliant
HTML5 tokenization and tree construction when parsing templates.
Using the DOM API to extract the results of tree construction
seems clunky, though.  Together with jsdom (a dependency) it's
a big blob of server-only code.

Turns out HTML5 tokenization and tree construction is not a small
spec.  Have to figure out how much support we need in templates
(compare with Ractive, e.g.).