Dont throw an error if weak package dont exist
This limits development of packages, specially testing packages that
depends on external/non-core packages.

Since its a weak dependency we would expect that the package survive
even if the weak dependency is not found.

I'm having an issue where I have a weak dependency - Its not installed
but I didnt expect that to be a problem.