Added runtime appcache
Reason:
We are working on improving the way cordova and meteor works together.
We want to use Meteor as a remote url in cordova so the Meteor server
has to serve device specific bundles to cordova.
Sometimes we have to serve sensitive information like the android push
api key - but at the moment the cordova uses the native webkit, so we
cannot determin if actually in cordova or just in the native browser.
----
Solution:
adding a querystring eg. `?cordova=3.0.0` would solve things but also
complicate stuff

I've built a runtime bundler allowing packages and the app to bundle
Assets depending on userAgent and querystring.

Now when on mobile its vital to be able to use the appcache - so this
code adds a hook for adding runtime bundles in the appcache while
preserving the querystring - unless hash is forced