Fix error on this.removed during session shutdown.
Followup to 4f4e5342 which removed session resumption. That commit
deferred the _deactivateAllSubscriptions call on session shutdown, but
left around the immediate deletion of the merge box
state (collectionViews). This meant that if any publish callbacks were
called before the _deactivateAllSubscriptions call, they would operate
on the invalid collectionViews state.

This change ensures that the subscriptions are considered deactivated as
soon as we decide to deactivate them, even before we actually deactivate
them (and call their possibly-yielding stop callbacks).