don't try to setAttribute 'type' for INPUT
IE8 doesn't let you set INPUT's type, even to the same value it already has.  When an element has reactive attributes, we set all the attributes when we get an invalidation, so there would be an error even if the type attribute was constant.  Now you only get an error if you actually try to change the type attribute from one thing to another.  In theory we could eventually support changing the type of an input by replacing the element (same as changing the contents of a STYLE element by replacing the element).