Two fixes to #each over arrays.
(a) Call observeSequence's changed callback with the correct order of
    arguments.

(b) In observeSequence's autorun, only be sensitive to reactive sources
    in the sequenceFunc. Previously, the entire autorun was run
    reactively, including its callbacks. This included the added
    callback in 'each', which meant that the UI.render calls for
    individual parts of the 'each' were run inside the autorun. This
    meant that when the autorun was invalidated, the autoruns inside
    UI.render would be stopped and the components would never update
    their DOM again. While this helped with the issue where UI.render
    doesn't stop its autoruns, it meant that the changed callbacks did
    nothing caused #1605.

These were only visible when your #each'd function returned an array,
not a cursor, and when the items in the array contained _id fields. If
it returned a cursor, the outer autorun didn't get invalidated on
changes to the contents of the cursor (just changes to the cursor
itself), and if the items didn't contain _id fields 'changed' could
never be called.

(Leaving #1605 open pending tests.)