Also include path of file in hash computation.
Followup to 72d3e08067. Without this, the hash wouldn't change and the client
wouldn't reload if a file was renamed in a way that didn't change the order
of files (eg, preserving its name in alphabetical order).

This didn't used to be a problem, as the hash computation was only used in
appcached, which included the file path already and would change if the
file was renamed.