Stop using handlebars in the bundler.
The dev bundle contains a copy of the handlebars NPM module solely for
creating app.html.  This is separate from the NPM module used by the
handlebars NPM package.

On the shark branch, we no longer use the handlebars NPM module for
Meteor template (it is being replaced by Spacebars), so in preparation
for that, we'll remove this barely-used build-time dependency on
handlebars.

A subsequent commit will remove it from the dev bundle.

Once the Spacebars API has fully settled (eg, it has been merged to
devel), we should get rid of this ad hoc templating and replace it with
Spacebars, either in webapp_server (driven entirely by program.json) or
by using unipackage.load in bundler.