Don't expect {{toString}} to be not found
We used to test that `{{toString}}` is undefined on a fresh
template, but now it resolves as a helper, along with
various other bits of name pollution that we can't tell
apart from real helpers/methods now that we search the
prototype chain, like `{{data}}`, `{{events}}`, `{{kind}}`,
and so on.  We should do something about that at some point.