Work around Node issue #6506.
We were previously doing this with a fork of Node in the dev bundle, but
that wouldn't help users in production.

We now expect users to use precisely v0.10.22 in production. When new
versions of Node come out, if users choose to use them before we get a
chance to put out a new Meteor, we disable websockets to work around the
most common effect of the issue instead of monkey-patching.