make .parented() work so #each can use it
The problem was we don't currently chain callbacks like rendered and parented when a class and superclass have different implementations.  The plumbing here is all rather arbitrary, considering our inheritance model is not finalized (and we may even do away with it, a la React) and "parented" is at best an undocumented but usable peer to "rendered".

Renamed the "parented" signal from DomRange to Component "notifyParented".