Make htmljs Tag constructor IE 8 safe
To make tags print nicely in the console, we dynamically create constructor functions of the form `function DIV() { ... }`, which is still the only thing Chrome responds to.  Modern browsers interpret a function expression of this form as exposing a symbol `DIV` only to lexical scope of the *body* of the function.  IE 8 does something else -- I forget what exactly, but it's not good.  Also I think `new Function("... code ...")` captures scope in IE 8, so we need to be careful what sort of names we inject.