meteor_npm: don't lose interesting resolved
If an indirect dependency resolves to something other than a semver (or
a GitHub tarball), it will be stored in the 'resolved'. Our shrinkwrap
minifier (which helps to reduce spurious shrinkwrap file changes) needs
to recognize that.

Also, consistently use the "version" field in the minified shrinkwrap
file (which a comment already claimed we could do).

Fixes #1684.